# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Commission Calculation System API
  version: "1.0.0"
  description: >
    REST API that powers the Commission Calculation System. The service exposes
    endpoints for authentication, agent management, sales capture, dashboards,
    reporting, and clawback operations.
servers:
  - url: http://localhost:5002
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SaleRequest:
      type: object
      properties:
        policy_number:
          type: string
          description: Existing policy number to reuse. If absent a new number is generated.
        product:
          type: string
          default: Life
        fyc_rate:
          type: number
          format: float
          default: 0.5
        premium:
          type: number
          format: float
        seller_id:
          type: integer
        sale_date:
          type: string
          format: date
        customer_name:
          type: string
        mobile:
          type: string
        gender:
          type: string
        notes:
          type: string
      required:
        - premium
        - seller_id
    SaleRecordResponse:
      type: object
      properties:
        message:
          type: string
        sale_id:
          type: integer
        policy_number:
          type: string
      required:
        - message
        - sale_id
        - policy_number
    SaleHistoryItem:
      type: object
      properties:
        id:
          type: integer
        policy_number:
          type: string
        sales_rep:
          type: string
        customer:
          type: string
        product:
          type: string
        amount:
          type: number
          format: float
        date:
          type: string
          format: date
    PaginatedSales:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleHistoryItem'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required:
        - items
        - page
        - limit
        - total
    PolicyLookup:
      type: object
      properties:
        exists:
          type: boolean
        policy:
          type: object
          nullable: true
          properties:
            number:
              type: string
            product:
              type: string
            fyc_rate:
              type: number
              format: float
        sales:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              customer:
                type: string
                nullable: true
              seller_id:
                type: integer
              premium:
                type: number
                format: float
              date:
                type: string
                format: date
      required:
        - exists
...
paths:
...
  /sales:
    post:
      tags:
        - Sales
      security:
        - bearerAuth: []
      summary: Record a sale and trigger commission calculations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleRequest'
      responses:
        '200':
          description: Sale recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleRecordResponse'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sales/next-policy-number:
    get:
      tags:
        - Sales
      security:
        - bearerAuth: []
      summary: Retrieve the next generated policy number
      responses:
        '200':
          description: Generated policy number
          content:
            application/json:
              schema:
                type: object
                properties:
                  next_policy_number:
                    type: string
  /sales/recent:
    get:
      tags:
        - Sales
      security:
        - bearerAuth: []
      summary: Latest 10 sales for dashboard widgets
      responses:
        '200':
          description: Recent sales
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    sales_rep:
                      type: string
                    customer:
                      type: string
                    amount:
                      type: number
                    date:
                      type: string
                      format: date
  /sales/history:
    get:
      tags:
        - Sales
      security:
        - bearerAuth: []
      summary: Paged sales history with simple text search
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: q
          schema:
            type: string
            description: Search filter against customer name, product, or mobile
      responses:
        '200':
          description: Paginated sales results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSales'
  /sales/policy-lookup:
    get:
      tags:
        - Sales
      security:
        - bearerAuth: []
      summary: Look up a policy number and its sales history
      parameters:
        - in: query
          name: policy_number
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lookup result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyLookup'
        '400':
          description: Missing policy number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...